refactor(hooks): split hook subsystem into focused modules
