+++ internal/hooks/runtime_config.go
@@ -1,0 +1,63 @@
+package hooks
+
+import (
+	"os"
+	"path/filepath"
+	"strconv"
+	"time"
+
+	"github.com/cristianoliveira/tmux-intray/internal/colors"
+	"github.com/cristianoliveira/tmux-intray/internal/config"
+)
+
+func isHooksVerbose() bool {
+	return os.Getenv("TMUX_INTRAY_HOOKS_VERBOSE") == "1"
+}
+
+func getHooksDir() string {
+	config.Load()
+	if dir := os.Getenv("TMUX_INTRAY_HOOKS_DIR"); dir != "" {
+		return dir
+	}
+	if dir := config.Get("hooks_dir", ""); dir != "" {
+		colors.Debug("hooks_dir from config: " + dir)
+		return dir
+	}
+	if configDir := os.Getenv("XDG_CONFIG_HOME"); configDir != "" {
+		return filepath.Join(configDir, "tmux-intray", "hooks")
+	}
+	home, _ := os.UserHomeDir()
+	return filepath.Join(home, ".config", "tmux-intray", "hooks")
+}
+
+func getFailureMode() string {
+	if mode := os.Getenv("TMUX_INTRAY_HOOKS_FAILURE_MODE"); mode != "" {
+		return mode
+	}
+	return "warn"
+}
+
+func getAsyncEnabled() bool {
+	if async := os.Getenv("TMUX_INTRAY_HOOKS_ASYNC"); async != "" {
+		return async == "1" || async == "true" || async == "yes" || async == "on"
+	}
+	return false
+}
+
+func getAsyncTimeout() time.Duration {
+	if timeoutStr := os.Getenv("TMUX_INTRAY_HOOKS_ASYNC_TIMEOUT"); timeoutStr != "" {
+		if seconds, err := time.ParseDuration(timeoutStr + "s"); err == nil {
+			return seconds
+		}
+	}
+	return 30 * time.Second
+}
+
+func getMaxAsyncHooks() int {
+	if maxStr := os.Getenv("TMUX_INTRAY_MAX_HOOKS"); maxStr != "" {
+		if max, err := strconv.Atoi(maxStr); err == nil && max > 0 {
+			return max
+		}
+	}
+	return 10
+}
